feat(UI): Added default banner option
Added a default banner option for generated applications. If yes is chosen then a common folder with screen will appear. And an import statement and self.intro setting will be added to app.py. If a banner is not chosen then these files changes will be erased using the post_gen cookiecutter hook. Also if the common folder is empty then it will be deleted as well.

closes #5